+++ app/models/category_result.rb
@@ -6,11 +6,9 @@ class CategoryResult < ApplicationRecord
   belongs_to :competition
   belongs_to :skater
   
-
+  ## scope
   scope :with_competition, ->{ joins(:competition) }
   scope :recent, ->{ with_competition.order("competitions.start_date desc") }
-  
-  scope :search_by_category, ->(cat) { where(category: cat) }
-
+  scope :category, ->(cat) { where(category: cat) }
   scope :top_rankers, ->(n) { where("ranking > 0 and ranking <= ? ", n.to_i).order(:ranking) }
 end