+++ CHANGELOG.md
@@ -1,12 +1,10 @@
 ## TODO
 - re-consider parsers/* system
 - http://www.figureskatingresults.fi/results/1415/CSFIN2014/ : charset="unicode" doesnt work
-- element.element, comoponent.component: name ?
 - app/views/score/show.html: unsafe youtube link
 - select options default sort
   - caching
-- category summary on competition as decorator
-- competition json: show. category summary requires attributes methods
+- jbuilder for .json
 
 ### new feature
 
@@ -14,6 +12,11 @@
 - sort in list table
 - eliminate feedback
 
+### check before release
+- app/controller/application_controller: unless Rails.env.develop
+- config/application: Version
+- config/database: /tmp
+- config/competitions.yml: debug
 
 ## 1.0.1
 - rake update:skaters..name space for update
@@ -36,6 +39,8 @@
 - elements/components-operators
 - updater/parser => skater/competition/score
 - lib/fisk8viewer => lib/fisk8_result_viewer
+- category summary on competition as decorator
+- element.element, comoponent.component: name as attr
 
 ## 1.0.0